remove cpp
